examples/app_examples/image_application/requirements.txt
lightning-flash[image,serve] == 0.7.0
streamlit_autorefresh
